Add footer links data file and updated file structure and the naming conventions
