+++ src/app/(frontend)/merchandise/_components/MerchandisePage.tsx
@@ -1,0 +1,11 @@
+import { MerchandiseFooterSection } from './footer/MerchandiseFooterSection'
+import { MerchandiseHeaderSection } from './header/MerchandiseHeaderSection'
+
+export function MerchandisePage() {
+  return (
+    <>
+      <MerchandiseHeaderSection />
+      <MerchandiseFooterSection />
+    </>
+  )
+}
