+++ src/app/(frontend)/merchandise/_components/footer/MerchandiseFooterData.tsx
@@ -1,0 +1,18 @@
+export const footerLinks = [
+  {
+    title: 'Canoe & Kayak North Shore',
+    href: 'https://canoeandkayak.co.nz/',
+  },
+  {
+    title: 'FurtherFasterNZ',
+    href: 'https://www.furtherfaster.co.nz/',
+  },
+  {
+    title: 'NRS',
+    href: 'https://www.nrs.com/',
+  },
+  {
+    title: 'Paddle Power',
+    href: 'https://paddle-power.nz/',
+  },
+]
