+++ src/app/(frontend)/merchandise/page.tsx
@@ -1,4 +1,4 @@
-import { MerchandisePage } from './_components/header/MerchandiseHeaderSection'
+import { MerchandisePage } from './_components/MerchandisePage'
 
 export default function MerchPage() {
   return <MerchandisePage />
